Conflicts:
	Frameworks/Core/ERExtensions/Sources/er/extensions/components/html5/ERXWOTelField.java
	Frameworks/Misc/ERPersistentSessionStorage/Resources/Properties
	Frameworks/Misc/ERQuartzScheduler/Components/ERQSJobInformations.wo/ERQSJobInformations.html
	Frameworks/Misc/ERQuartzScheduler/Resources/Properties
	Tests/ERXTest/Sources/er/extensions/crypting/TestBCrypt.java

+++ Tests/ERXTest/Sources/er/extensions/eof/ERXKeyTest.java
@@ -20,7 +20,9 @@ import er.erxtest.model.Company;
 import er.erxtest.model.Employee;
 import er.erxtest.model.Paycheck;
 import er.erxtest.model.Role;
+
 import er.extensions.foundation.ERXArrayUtilities;
+import er.extensions.eof.ERXEOAccessUtilities;
 
 public class ERXKeyTest extends ERXTestCase {
 
@@ -139,37 +141,52 @@ public class ERXKeyTest extends ERXTestCase {
 		shinraNonManager = null;
 		shinraNonManagerNames = null;
 		tyrellEmployees = null;
+
 		paychecks = null;
 
-		acme.delete();
-		shinraInc.delete();
-		tyrellCorp.delete();
-		sortCo.delete();
+		if (ERXTestCase.adaptorName().equals("JDBC")) {
+
+			ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "Company"), "delete from Company");
+			ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "Employee"), "delete from Employee");
+                	ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "EmployeeHI"), "delete from EmployeeHI");
+                	ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "EmployeeVI"), "delete from EmployeeVI");
+                	ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "EmployeeRole"), "delete from EmployeeRole");
+			ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "Paycheck"), "delete from Paycheck");
+			ERXEOAccessUtilities.evaluateSQLWithEntity(ec, ERXEOAccessUtilities.entityNamed(ec, "Role"), "delete from Role");
+
+		}
+
+		if (ERXTestCase.adaptorName().equals("Memory")) {
+
+			shinraInc.delete();
+			tyrellCorp.delete();
+			sortCo.delete();
 		
-		moe.delete();
-		larry.delete();
-		curly.delete();
-		tyrell.delete();
-		deckard.delete();
-		rachael.delete();
-		rufus.delete();
-		rude.delete();
-		elena.delete();
-		sephiroth.delete();
-		aaBb.delete();
-		Abba.delete();
+			moe.delete();
+			larry.delete();
+			curly.delete();
+			tyrell.delete();
+			deckard.delete();
+			rachael.delete();
+			rufus.delete();
+			rude.delete();
+			elena.delete();
+			sephiroth.delete();
+			aaBb.delete();
+			Abba.delete();
+
+			paycheck5000.delete();
+			paycheck10000.delete();
+			paycheck15000.delete();
 		
-		paycheck5000.delete();
-		paycheck10000.delete();
-		paycheck15000.delete();
-		
-		headHunter.delete();
-		bigBoss.delete();
-		accountant.delete();
-		doctor.delete();
-		barber.delete();
-		
-		ec.saveChanges();
+			headHunter.delete();
+			bigBoss.delete();
+			accountant.delete();
+			doctor.delete();
+			barber.delete();
+
+			ec.saveChanges();
+		}
 	}
 
 	@Test
